Create prime_armstrong.java
This module gives insight on prime numbers and armstrong numbers. The first half program computes and lists all prime numbers between 1-100 and the second half prompts for a number to check if it is armstrong or not.

Prime Number: Only two factors - 1 and the number itself. No other factor. Ex: 2,3,5,7,11.etc

Armstrong Number: The sum of each digits to the power of total digits which equates to the number itself. 
Ex: 123 => total digits = 3
                  1^3 + 2^3 + 3^3 = 123